requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
# Примечание: версии указаны как пример, можно обновлять при необходимости
